Fixed Bug: Volumes were not correctly unmounted
Fixed Bug: Volumes were not correctly unmounted
Description: Linux and Mac use a command to unmount and override the default umount-method defined by AbstractFuseFS to achieve that. jnr-fuse requires a call to a native lib to close a filesystem instead, which is usally located in AbstractFuseFS#umount. To allow access to both approaches without reflection or a breach of encapsulation the bypass system was introduced.